pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["JimZer <jimi.vaubien@protonmail.com>"]
description = "A CLI tool to help lazy coders get the work done with AI (commit messages, pull requests ...)"
homepage = "https://github.com/bitswired/lazycodr"
name = "lazycodr"
packages = [{include = "lazycodr", from = "src"}]
readme = "README.md"
repository = "https://github.com/bitswired/lazycodr"
version = "0.1.2"


[tool.poetry.dependencies]
GitPython = "^3.1.37"
httpx = "^0.24.0"
jinja2 = "^3.1.2"
langchain = "^0.0.312"
openai = "^0.27.6"
pydantic = "^1.10.7"
pyfiglet = "^0.8.post1"
pygithub = "^1.58.2"
python = "^3.10"
rich = "^13.3.5"
tiktoken = "^0.4.0"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.3.1"
pytest = "^7.3.1"

[tool.poetry.scripts]
lazycodr = "lazycodr.cli:main"
